==> src/Competencia2/05_NA_limpieza_y_reparacion.ipynb <==
#limpio la memoria
rm( list=ls() )  #remove all objects
gc()             #garbage collection

require("data.table")

#setwd("C:/_MCD/Labo1/code")  #Establezco el Working Directory
setwd("~/buckets/b1/")

#cargo el dataset
dataset  <- fread("./datasets/competencia2_2022.csv.gz")   #donde entreno

nrow(dataset)

CorregirCampoMes  <- function( pcampo, pmeses )
{
  tbl <- dataset[  ,  list( "v1" = shift( get(pcampo), 1, type="lag" ),
                            "v2" = shift( get(pcampo), 1, type="lead" )
  ), 
  by=numero_de_cliente ]
  
  tbl[ , numero_de_cliente := NULL ]
  tbl[ , promedio := rowMeans( tbl,  na.rm=TRUE ) ]
  
  dataset[ ,
           paste0(pcampo) := ifelse( !(foto_mes %in% pmeses),
                                     get( pcampo),
                                     tbl$promedio ) ]
}

Corregir_MachineLearning  <- function( dataset )
{
  gc()
  #acomodo los errores del dataset
##  
  dataset[ foto_mes==202006,  active_quarter   := NA ]

##
  dataset[ foto_mes==201905,  mrentabilidad     := NA ]
  dataset[ foto_mes==201905,  mrentabilidad_annual     := NA ]
  dataset[ foto_mes==201905,  mcomisiones      := NA ]
  dataset[ foto_mes==201905,  mpasivos_margen  := NA ]
  dataset[ foto_mes==201905,  mactivos_margen  := NA ]  
    
  dataset[ foto_mes==201910,  mrentabilidad     := NA ]
  dataset[ foto_mes==201910,  mrentabilidad_annual        := NA ]
  dataset[ foto_mes==201910,  mcomisiones       := NA ]
  dataset[ foto_mes==201910,  mactivos_margen   := NA ]
  dataset[ foto_mes==201910,  mpasivos_margen   := NA ]    
   
  dataset[ foto_mes==202006,  mrentabilidad   := NA ]
  dataset[ foto_mes==202006,  mrentabilidad_annual   := NA ]
  dataset[ foto_mes==202006,  mcomisiones   := NA ]
  dataset[ foto_mes==202006,  mactivos_margen   := NA ]
  dataset[ foto_mes==202006,  mpasivos_margen   := NA ]    
##    
  dataset[ foto_mes==202006,  mcuentas_saldo   := NA ]
##
  dataset[ foto_mes==202006,  ctarjeta_debito_transacciones  := NA ]
  dataset[ foto_mes==202006,  mautoservicio   := NA ]
##
  dataset[ foto_mes==202006,  ctarjeta_visa_transacciones   := NA ]
  dataset[ foto_mes==202006,  mtarjeta_visa_consumo   := NA ]
##
  dataset[ foto_mes==202006,  ctarjeta_master_transacciones   := NA ]
  dataset[ foto_mes==202006,  mtarjeta_master_consumo   := NA ]
##
  dataset[ foto_mes==201904,  ctarjeta_visa_debitos_automaticos  :=  NA ]
  dataset[ foto_mes==201904,  mttarjeta_visa_debitos_automaticos := NA ]    
##    
  dataset[ foto_mes==201901,  cpagodeservicios := NA ]    
  dataset[ foto_mes==201902,  cpagodeservicios := NA ]    
  dataset[ foto_mes==201903,  cpagodeservicios := NA ]    
  dataset[ foto_mes==201904,  cpagodeservicios := NA ]    
  dataset[ foto_mes==201905,  cpagodeservicios := NA ]    
  dataset[ foto_mes==201901,  mpagodeservicios := NA ]    
  dataset[ foto_mes==201902,  mpagodeservicios := NA ]    
  dataset[ foto_mes==201903,  mpagodeservicios := NA ]    
  dataset[ foto_mes==201904,  mpagodeservicios := NA ]    
  dataset[ foto_mes==201905,  mpagodeservicios := NA ]     
  dataset[ foto_mes==201901,  cpagomiscuentas := NA ]    
  dataset[ foto_mes==201902,  cpagomiscuentas := NA ]    
  dataset[ foto_mes==201903,  cpagomiscuentas := NA ]    
  dataset[ foto_mes==201904,  cpagomiscuentas := NA ]    
  dataset[ foto_mes==201905,  cpagomiscuentas := NA ]     
  dataset[ foto_mes==201901,  mpagomiscuentas := NA ]    
  dataset[ foto_mes==201902,  mpagomiscuentas := NA ]    
  dataset[ foto_mes==201903,  mpagomiscuentas := NA ]    
  dataset[ foto_mes==201904,  mpagomiscuentas := NA ]    
  dataset[ foto_mes==201905,  mpagomiscuentas := NA ]        
##    
  dataset[ foto_mes==201901,  ccomisiones_mantenimiento  := NA ]
  dataset[ foto_mes==201901,  mcomisiones_mantenimiento  := NA ]    
  dataset[ foto_mes==201905,  ccomisiones_otras := NA ]
  dataset[ foto_mes==201910,  ccomisiones_otras := NA ]
  dataset[ foto_mes==202006,  ccomisiones_otras   := NA ]    
  dataset[ foto_mes==201905,  mcomisiones_otras := NA ]    
  dataset[ foto_mes==201910,  mcomisiones_otras := NA ]    
  dataset[ foto_mes==202006,  mcomisiones_otras   := NA ]    
##    
  dataset[ foto_mes==201901,  ctransferencias_recibidas  := NA ]
  dataset[ foto_mes==201902,  ctransferencias_recibidas  := NA ]
  dataset[ foto_mes==201903,  ctransferencias_recibidas  := NA ]
  dataset[ foto_mes==201904,  ctransferencias_recibidas  := NA ]
  dataset[ foto_mes==201905,  ctransferencias_recibidas  := NA ]
  dataset[ foto_mes==201901,  mtransferencias_recibidas  := NA ]
  dataset[ foto_mes==201902,  mtransferencias_recibidas  := NA ]
  dataset[ foto_mes==201903,  mtransferencias_recibidas  := NA ]
  dataset[ foto_mes==201904,  mtransferencias_recibidas  := NA ]
  dataset[ foto_mes==201905,  mtransferencias_recibidas  := NA ]
#
 dataset[ foto_mes==202006,  cextraccion_autoservicio   := NA ]
  dataset[ foto_mes==202006,  mextraccion_autoservicio   := NA ]
  dataset[ foto_mes==202006,  ccheques_depositados   := NA ]
  dataset[ foto_mes==202006,  mcheques_depositados   := NA ]
  dataset[ foto_mes==202006,  ccheques_emitidos   := NA ]
  dataset[ foto_mes==202006,  mcheques_emitidos   := NA ]
  dataset[ foto_mes==202006,  ccheques_depositados_rechazados   := NA ]
  dataset[ foto_mes==202006,  mcheques_depositados_rechazados   := NA ]
  dataset[ foto_mes==202006,  ccheques_emitidos_rechazados   := NA ]
  dataset[ foto_mes==202006,  mcheques_emitidos_rechazados   := NA ]
  dataset[ foto_mes==202006,  tcallcenter   := NA ]
  dataset[ foto_mes==202006,  ccallcenter_transacciones   := NA ]
  dataset[ foto_mes==202006,  thomebanking   := NA ]
  dataset[ foto_mes==201910,  chomebanking_transacciones   := NA ]    
  dataset[ foto_mes==202006,  chomebanking_transacciones   := NA ]
  dataset[ foto_mes==202006,  ccajas_transacciones   := NA ]
  dataset[ foto_mes==202006,  ccajas_consultas   := NA ]
  dataset[ foto_mes==202006,  ccajas_depositos   := NA ]
  dataset[ foto_mes==202006,  ccajas_extracciones   := NA ]
  dataset[ foto_mes==202006,  ccajas_otras   := NA ]
  dataset[ foto_mes==202006,  catm_trx   := NA ]
  dataset[ foto_mes==202006,  matm   := NA ]
  dataset[ foto_mes==202006,  catm_trx_other   := NA ]
  dataset[ foto_mes==202006,  matm_other   := NA ]      
##    
  dataset[ foto_mes==201904,  Master_mfinanciacion_limite  := NA ]
  dataset[ foto_mes==202104,  Master_mfinanciacion_limite  := NA ]
##    
  dataset[ foto_mes==201907,  Master_fultimo_cierre  := NA ]
  dataset[ foto_mes==202009,  Master_fultimo_cierre  := NA ]
##  
  dataset[ foto_mes==202012,  Master_mpagado  := NA ]
##    
  dataset[ foto_mes==201904,  Visa_mfinanciacion_limite  := NA ]
##    
  dataset[ foto_mes==201907,  Visa_fultimo_cierre  := NA ]
  dataset[ foto_mes==202009,  Visa_fultimo_cierre  := NA ]
##  
  dataset[ foto_mes==202012,  Visa_mpagado  := NA ]
}

#Elimino los campos problematicos
dataset[  , internet := NULL ]
dataset[  , tmobile_app := NULL ]
dataset[  , cmobile_app_trx := NULL ]


#Ordeno para imputar correctamente
setorder( dataset, numero_de_cliente, foto_mes )

Corregir_MachineLearning( dataset )

fwrite( dataset,
        file="./datasets/competencia2_2022_reparacNA.csv.gz",
        logical01= TRUE,
        sep= "," )

dataset[numero_de_cliente %in% c(35293777),list(numero_de_cliente,foto_mes,mrentabilidad)]
