==> src/Competencia2/05_nas_ratio_limp1.ipynb <==
rm( list=ls() )  #remove all objects
gc()             #garbage collection

require("data.table")

setwd("C:/_MCD/Labo1/code")  #Establezco el Working Directory

#cargo el dataset
dataset  <- fread("./datasets/competencia2_2022_limpieza1.csv.gz")   #donde entreno

#creo la carpeta donde va el experimento
# DA  representa  Exploratory Data Analysis
dir.create( "./exp/EDA-comp2/", showWarnings = FALSE )
setwd("./exp/EDA-comp2/")   #Establezco el Working Directory DEL EXPERIMENTO

#ordeno el dataset
setorder( dataset,  foto_mes )

campos_buenos  <- setdiff( colnames( dataset), 
                           c("numero_de_cliente","foto_mes","clase_ternaria" ) )



pdf("nas_ratio_limpieza1.pdf")

for( campo in  campos_buenos )
{
  tbl <- dataset[ foto_mes<=202105,
                  list( "na_ratio" = sum( is.na(get(campo)), na.rm=TRUE)/.N ) , 
                  foto_mes ]

  #ymin <-  min( tbl$na_ratio )
  #ymax <-  max( tbl$na_ratio )
       ymin <- 0
    ymax <- 1
  if( ymin == 0 )  ymin <- -0.1
  if( ymax == 0 )  ymax <-  0.1

  plot(x= 1:nrow(tbl),
       y= tbl$na_ratio,
       type= "o",
       main= paste0("NAs ratio  -  ",  campo),
       xlab= "Periodo",
       ylab= "NAs  ratio",
       ylim= c( ymin, ymax ),
       xaxt= "n"
     )

  axis(1, at=1:nrow(tbl), labels=tbl$foto_mes)

   abline(v=c(1,12), col=c("green","green","green"), lty=c(1,1,1), lwd=c(1,1,1))
}

dev.off()

#dejo la marca final
cat( format(Sys.time(), "%Y%m%d %H%M%S"),"\n",
     file= "zRend.txt",
     append= TRUE  )
