==> src/Competencia2/09-EC_analisis_drifting.ipynb <==
#limpio la memoria
rm( list=ls() )  #remove all objects
gc()             #garbage collection

require("data.table")
require("scales")
require("lightgbm")

#Parametros del script
#setwd("C:/_MCD/Labo1/code")  #Establezco el Working Directory
setwd("~/buckets/b1/")

#cargo el dataset
dataset  <- fread("./datasets/competencia2_2022_reparacEstad.csv.gz")   #donde entreno# FIN Parametros del script

kexperimento <- "comp2-EDA"
#creo la carpeta donde va el experimento
dir.create( paste0( "./exp/", kexperimento, "/"), showWarnings = FALSE )
setwd(paste0( "./exp/", kexperimento, "/"))   #Establezco el Working Directory DEL EXPERIMENTO


#creo la clase_binaria SI={ BAJA+1, BAJA+2 }    NO={ CONTINUA }
dataset[ foto_mes<=202103, 
         clase01 :=  ifelse( clase_ternaria=="CONTINUA", 0, 1 ) ]

dataset[  , foto_mes := as.character( foto_mes ) ]

campos_ordenados <- c("ctrx_quarter","cpayroll_trx","mcaja_ahorro","mpayroll","mcuentas_saldo","mpasivos_margen","mtarjeta_visa_consumo","mprestamos_personales","mrentabilidad_annual","mcuenta_corriente","mactivos_margen","active_quarter","ctarjeta_visa_transacciones","cprestamos_personales","Visa_fechaalta","mcomisiones_mantenimiento","cdescubierto_preacordado","Visa_status","Visa_mpagominimo","Master_fechaalta","cliente_edad","mrentabilidad","Master_Fvencimiento","Visa_msaldototal","numero_de_cliente","Visa_Fvencimiento","cproductos","cliente_antiguedad","ctarjeta_master","Visa_msaldopesos","ccaja_seguridad","Master_mfinanciacion_limite","mcuenta_debitos_automaticos","Master_status","mtransferencias_recibidas","chomebanking_transacciones","ccomisiones_mantenimiento","mautoservicio","ccajas_consultas","mtarjeta_master_consumo","ccomisiones_otras","Visa_cconsumos","Visa_mfinanciacion_limite","mcaja_ahorro_dolares","mttarjeta_visa_debitos_automaticos","mcomisiones_otras","mcomisiones","Visa_mpagospesos","Visa_mconsumospesos","Master_mlimitecompra","ccallcenter_transacciones","thomebanking","ccaja_ahorro","tcallcenter","Visa_mlimitecompra","ctransferencias_recibidas","Master_msaldototal","Master_fultimo_cierre","mtransferencias_emitidas","ctarjeta_debito_transacciones","mpagomiscuentas","Visa_fultimo_cierre","ctarjeta_visa_debitos_automaticos","Master_msaldopesos","Visa_mconsumosdolares","ctarjeta_visa","ctarjeta_master_transacciones","Visa_mconsumototal","Master_mconsumospesos","mplazo_fijo_dolares","ctarjeta_debito","Master_mpagominimo","Master_mpagospesos","ccuenta_debitos_automaticos","ctransferencias_emitidas","mextraccion_autoservicio","Visa_mpagado","ccajas_transacciones","cextraccion_autoservicio","cpagomiscuentas","Master_cconsumos","Visa_msaldodolares","matm_other","Master_mconsumototal","matm","ccajas_otras","catm_trx","cseguro_vida","ccajas_depositos","ccajas_extracciones","cinversion1","mcheques_depositados","mttarjeta_master_debitos_automaticos","Visa_mpagosdolares","cplazo_fijo","Master_madelantopesos","Master_mpagado","cseguro_accidentes_personales","mforex_sell","catm_trx_other","Master_mpagosdolares","Master_mconsumosdolares","cseguro_vivienda","mcaja_ahorro_adicional","Visa_Finiciomora","minversion2","cinversion2","ctarjeta_master_debitos_automaticos","ccheques_emitidos","Visa_delinquency","Master_msaldodolares","mcheques_emitidos","Master_cadelantosefectivo","Master_delinquency","ccheques_depositados","Visa_madelantopesos","mtarjeta_visa_descuentos","Visa_cadelantosefectivo","Visa_madelantodolares","cforex","cliente_vip","tcuentas","ccuenta_corriente","mcuenta_corriente_adicional","cprestamos_prendarios","mprestamos_prendarios","cprestamos_hipotecarios","mprestamos_hipotecarios","mplazo_fijo_pesos","minversion1_pesos","minversion1_dolares","cseguro_auto","mpayroll2","cpayroll2_trx","cpagodeservicios","mpagodeservicios","ccajeros_propios_descuentos","mcajeros_propios_descuentos","ctarjeta_visa_descuentos","ctarjeta_master_descuentos","mtarjeta_master_descuentos","cforex_buy","mforex_buy","cforex_sell","ccheques_depositados_rechazados","mcheques_depositados_rechazados","ccheques_emitidos_rechazados","mcheques_emitidos_rechazados","Master_Finiciomora","Master_madelantodolares")

v_2019  <- c(201901,201902,201903,201904,201905,201906,201907,201908,201909,201910,201911,201912)
v_2020  <- c(202001,202002,202003,202004,202005,202006,202007,202008,202009,202010,202011,202012)
v_2021  <- c(202101, 202102, 202103, 202105)

GLOBAL_colores <-  viridis_pal()(length( periodos_analisis ) )

graficar_campo  <- function( campo, periodos_analisis )
{

  #quito de grafico las colas del 5% de las densidades

  tb_quantiles  <- data.table( qmin= numeric(),  qmax= numeric() )

  for( periodo in periodos_analisis )
  {
    qu  <- quantile(  dataset[ foto_mes==periodo , get(campo) ] , prob= c(0.05, 0.95), na.rm=TRUE )
    tb_quantiles  <- rbind( tb_quantiles, use.names=FALSE, as.list(qu) )
  }

  xxmin  <- tb_quantiles[ , min( qmin ) ]
  xxmax  <- tb_quantiles[ , max( qmax ) ]

  yymax  <- 0
  for( per in periodos_analisis )
  {
    den  <- density( dataset[ foto_mes==per, get(campo) ],
                     kernel="gaussian", na.rm=TRUE )

    mayor  <- max( den$y )
    if( mayor > yymax ) yymax <- mayor 
  }

  densidad_A  <- density( dataset[ foto_mes==periodos_analisis[1], get(campo) ],
                          kernel="gaussian", na.rm=TRUE )


  plot( densidad_A,
        col= GLOBAL_colores[1],
        xlim= c( xxmin, xxmax ),
        ylim= c( 0, yymax ),
        main= paste0( campo  ) 
      )

  for( per in 2:length(periodos_analisis) )
  {
    densidad_B  <- density( dataset[ foto_mes==periodos_analisis[ per ], get(campo) ],
                            kernel="gaussian", na.rm=TRUE )

    lines(densidad_B, col= GLOBAL_colores[per], lty=1)
  }
  
  legend(  "topright",  
           legend= periodos_analisis,
           col=GLOBAL_colores, lty=1 )

}

pdf("densidades_reparacEC_todo.pdf")

for( campo in  campos_ordenados )
{
  cat( campo, "  " )

  graficar_campo( campo, periodos_analisis )
}

dev.off()

v_2019  <- c(201901,201902,201903,201904,201905,201906,201907,201908,201909,201910,201911,201912)
v_2020  <- c(202001,202002,202003,202004,202005,202006,202007,202008,202009,202010,202011,202012)
v_2021  <- c(202101, 202102, 202103, 202105)

GLOBAL_colores <-  viridis_pal()(length( periodos_analisis ) )

pdf("densidades_reparacEC_2020y2021.pdf")

for( campo in  campos_ordenados )
{
  cat( campo, "  " )

  graficar_campo( campo, append(v_2020,v_2021) )
}

dev.off()

pdf("densidades_reparacEC_2021.pdf")

for( campo in  campos_ordenados )
{
  cat( campo, "  " )

  graficar_campo( campo, v_2021 )
}

dev.off()

pdf("densidades_reparacEC_desde202007.pdf")

for( campo in  campos_ordenados )
{
  cat( campo, "  " )

  graficar_campo( campo, append(c(202007,202008,202009,202010,202011,202012),v_2021) )
}

dev.off()
