==> src/Competencia2/9_03_limpieza_y_reparacion.ipynb <==
#limpio la memoria
rm( list=ls() )  #remove all objects
gc()             #garbage collection

require("data.table")

setwd("C:/_MCD/Labo1/code")  #Establezco el Working Directory

#cargo el dataset
dataset  <- fread("./datasets/competencia2_2022.csv.gz")   #donde entreno


nrow(dataset)

CorregirCampoMes  <- function( pcampo, pmeses )
{
  tbl <- dataset[  ,  list( "v1" = shift( get(pcampo), 1, type="lag" ),
                            "v2" = shift( get(pcampo), 1, type="lead" )
  ), 
  by=numero_de_cliente ]
  
  tbl[ , numero_de_cliente := NULL ]
  tbl[ , promedio := rowMeans( tbl,  na.rm=TRUE ) ]
  
  dataset[ ,
           paste0(pcampo) := ifelse( !(foto_mes %in% pmeses),
                                     get( pcampo),
                                     tbl$promedio ) ]
}

#------------------------------------------------------------------------------
# reemplaza cada variable ROTA  (variable, foto_mes)  con el promedio entre  ( mes_anterior, mes_posterior )

Corregir_EstadisticaClasica  <- function( dataset )
{
  CorregirCampoMes( "active_quarter", c(202006) )
  CorregirCampoMes( "mrentabilidad", c(201905, 201910, 202006) )
  CorregirCampoMes( "mrentabilidad_annual", c(201905, 201910, 202006) )
  CorregirCampoMes( "mcomisiones", c(201905, 201910, 202006) )
  CorregirCampoMes( "mpasivos_margen", c(201905, 201910, 202006) )
  CorregirCampoMes( "mactivos_margen", c(201905, 201910, 202006) )
  CorregirCampoMes( "mcuentas_saldo", c(202006) )
  CorregirCampoMes( "ctarjeta_debito_transacciones", c(202006) )
  CorregirCampoMes( "mautoservicio", c(202006) )
  CorregirCampoMes( "ctarjeta_visa_transacciones", c(202006) )
  CorregirCampoMes( "mtarjeta_visa_consumo", c(202006) )
  CorregirCampoMes( "ctarjeta_master_transacciones", c(202006) )
  CorregirCampoMes( "mtarjeta_master_consumo", c(202006) )
  CorregirCampoMes( "cprestamos_prendarios", c(202005) )
  CorregirCampoMes( "mprestamos_prendarios", c(202005) )
  CorregirCampoMes( "cprestamos_hipotecarios", c(202005) )    
  CorregirCampoMes( "mprestamos_hipotecarios", c(202005) )
  CorregirCampoMes( "mpayroll2", c(201901) )
  CorregirCampoMes( "cpayroll2_trx", c(201901) )
  CorregirCampoMes( "ctarjeta_visa_debitos_automaticos", c(201904) )
  CorregirCampoMes( "mttarjeta_visa_debitos_automaticos", c(201904) )
  CorregirCampoMes( "cpagodeservicios", c(201901,201902,201903,201904,201905) ) #revisar
  CorregirCampoMes( "mpagodeservicios", c(201901,201902,201903,201904,201905) ) #revisar
  CorregirCampoMes( "cpagomiscuentas", c(201901,201902,201903,201904,201905) ) #revisar
  CorregirCampoMes( "mpagomiscuentas", c(201901,201902,201903,201904,201905) ) #revisar
  CorregirCampoMes( "ccajeros_propios_descuentos", c(201902) )
  CorregirCampoMes( "mcajeros_propios_descuentos", c(201902) )
  CorregirCampoMes( "ccomisiones_mantenimiento", c(201901) )
  CorregirCampoMes( "mcomisiones_mantenimiento", c(201901) )
  CorregirCampoMes( "ccomisiones_otras", c(201905, 201910, 202006) )
  CorregirCampoMes( "mcomisiones_otras", c(201905, 201910, 202006) )  
  CorregirCampoMes( "ctransferencias_recibidas", c(201901,201902,201903,201904,201905) ) #revisar
  CorregirCampoMes( "mtransferencias_recibidas", c(201901,201902,201903,201904,201905) ) #revisar
  CorregirCampoMes( "cextraccion_autoservicio", c(202006) )
  CorregirCampoMes( "mextraccion_autoservicio", c(202006) )
  CorregirCampoMes( "ccheques_depositados", c(202006) )
  CorregirCampoMes( "mcheques_depositados", c(202006) )
  CorregirCampoMes( "mcheques_emitidos", c(202006) )
  CorregirCampoMes( "mcheques_emitidos", c(202006) )
  CorregirCampoMes( "ccheques_depositados_rechazados", c(202006) )
  CorregirCampoMes( "mcheques_depositados_rechazados", c(202006) )
  CorregirCampoMes( "ccheques_emitidos_rechazados", c(202006) )
  CorregirCampoMes( "mcheques_emitidos_rechazados", c(202006) )   
  CorregirCampoMes( "tcallcenter", c(202006) )
  CorregirCampoMes( "ccallcenter_transacciones", c(202006) )    
  CorregirCampoMes( "thomebanking", c(202006) )
  CorregirCampoMes( "chomebanking_transacciones", c( 201910, 202006) )
  CorregirCampoMes( "ccajas_transacciones", c(202006) )
  CorregirCampoMes( "ccajas_consultas", c(202006) )
  CorregirCampoMes( "ccajas_depositos", c(202006) )
  CorregirCampoMes( "ccajas_extracciones", c(202006) )
  CorregirCampoMes( "ccajas_otras", c(202006) )
  CorregirCampoMes( "catm_trx", c(202006) )
  CorregirCampoMes( "matm", c(202006) )
  CorregirCampoMes( "catm_trx_other", c(202006) )
  CorregirCampoMes( "matm_other", c(202006) )
  CorregirCampoMes( "ctrx_quarter", c(202006) )
  CorregirCampoMes( "cmobile_app_trx", c(201901,201902,201903,201904,201905,201906,202006) ) #revisar      
  CorregirCampoMes( "Visa_mfinanciacion_limite", c(202006) )
  CorregirCampoMes( "Master_mfinanciacion_limite", c(202006) )
  CorregirCampoMes( "Master_fultimo_cierre", c(201907,202009) )
  CorregirCampoMes( "Visa_fultimo_cierre", c(201907,202009) )
  CorregirCampoMes( "Master_mpagado", c(202012) )
  CorregirCampoMes( "Visa_mpagado", c(202012) )
#  CorregirCampoMes( "Visa_status", c(202010) ) #no se puede imputar por promedio


    


}

#Elimino los campos problematicos
#Internet se daño a partir de 202010
dataset[  , internet := NULL ]

#Internet se daño a partir de 202010
dataset[  , tmobile_app := NULL ]

#Ordeno para imputar correctamente
setorder( dataset, numero_de_cliente, foto_mes )

Corregir_EstadisticaClasica( dataset )

fwrite( dataset,
        file="./datasets/competencia2_2022_limpieza1.csv.gz",
        logical01= TRUE,
        sep= "," )

dataset[numero_de_cliente %in% c(35293777),list(numero_de_cliente,foto_mes,ctrx_quarter)]
