==> src/Competencia2/99_chequeoDatasets.ipynb <==
#limpio la memoria
rm( list=ls() )  #remove all objects
gc()             #garbage collection

require("data.table")

# Inicial

#Parametros del script
#setwd("C:/_MCD/Labo1/code")  #Establezco el Working Directory
setwd("~/buckets/b1/")

#cargo el dataset
dataset  <- fread("./datasets/competencia2_2022.csv.gz")   #donde entreno# FIN Parametros del script

nrow(dataset)

ncol(dataset)

# Reparación

#Parametros del script
#setwd("C:/_MCD/Labo1/code")  #Establezco el Working Directory
setwd("~/buckets/b1/")

#cargo el dataset
dataset  <- fread("./datasets/competencia2_2022_reparacEstad.csv.gz")   #donde entreno# FIN Parametros del script

nrow(dataset)

ncol(dataset)

#Parametros del script
#setwd("C:/_MCD/Labo1/code")  #Establezco el Working Directory
setwd("~/buckets/b1/")

#cargo el dataset
dataset  <- fread("./datasets/competencia2_2022_reparacNA.csv.gz")   #donde entreno# FIN Parametros del script

nrow(dataset)

ncol(dataset)

# FE Inicial

#Parametros del script
#setwd("C:/_MCD/Labo1/code")  #Establezco el Working Directory
setwd("~/buckets/b1/")

#cargo el dataset
dataset  <- fread("./datasets/c2_reparacEstad_fe_inicial.csv.gz")   #donde entreno# FIN Parametros del script

nrow(dataset)

ncol(dataset)

#Parametros del script
#setwd("C:/_MCD/Labo1/code")  #Establezco el Working Directory
setwd("~/buckets/b1/")

#cargo el dataset
dataset  <- fread("./datasets/c2_reparacNA_fe_inicial.csv.gz")   #donde entreno# FIN Parametros del script

nrow(dataset)
ncol(dataset)

# Drifting

#limpio la memoria
rm( list=ls() )  #remove all objects
gc()             #garbage collection

require("data.table")

#Parametros del script
#setwd("C:/_MCD/Labo1/code")  #Establezco el Working Directory
setwd("~/buckets/b1/")

#cargo el dataset
dataset  <- fread("./datasets/c2_reparacEstad_fe_inicial_rank0simple.csv.gz")   #donde entreno# FIN Parametros del script

nrow(dataset)
ncol(dataset)

#Parametros del script
#setwd("C:/_MCD/Labo1/code")  #Establezco el Working Directory
setwd("~/buckets/b1/")

#cargo el dataset
dataset  <- fread("./datasets/c2_reparacEstad_fe_inicial_rank0fijo.csv.gz")   #donde entreno# FIN Parametros del script

nrow(dataset)
ncol(dataset)

#Parametros del script
#setwd("C:/_MCD/Labo1/code")  #Establezco el Working Directory
setwd("~/buckets/b1/")

#cargo el dataset
dataset  <- fread("./datasets/c2_reparacEstad_fe_inicial_deflacion.csv.gz")   #donde entreno# FIN Parametros del script

# AD HOC

rm( list=ls() )  #remove all objects
gc()             #garbage collection

require("data.table")
setwd("~/buckets/b1/")


dataset  <- fread("./datasets/c2_reparacNA_fe_inicialylag1_deflacion.csv.gz")   #donde entreno# FIN Parametros del script

nrow(dataset)
ncol(dataset)

head(dataset)

colnames(dataset)
