==> src/Competencia2/99_promedio.ipynb <==
rm( list=ls() )  #remove all objects
gc()             #garbage collection

require("data.table")

#setwd("C:/_MCD/Labo1/code")  #Establezco el Working Directory
setwd("~/buckets/b1/")

#cargo el dataset
dataset  <- fread("./datasets/competencia2_2022.csv.gz")   #donde entreno

#creo la carpeta donde va el experimento
# DA  representa  Exploratory Data Analysis
dir.create( "./exp/comp2-EDA/", showWarnings = FALSE )
setwd("./exp/comp2-EDA/")   #Establezco el Working Directory DEL EXPERIMENTO

#ordeno el dataset
setorder( dataset,  foto_mes )

campos_buenos <- c("ctrx_quarter","mcuentas_saldo","mcaja_ahorro", "mcuenta_corriente","cpayroll_trx","mpayroll","mprestamos_personales","cprestamos_personales","mpasivos_margen","mrentabilidad","mrentabilidad_annual","mcomisiones_mantenimiento","mactivos_margen","cliente_edad","mtarjeta_visa_consumo","ctarjeta_visa_transacciones","Master_Fvencimiento","Master_fechaalta","Visa_Fvencimiento","Visa_fechaalta","ctarjeta_visa_transacciones")

pdf("promedios_desde202010.pdf")

for( campo in  campos_buenos )
{
  tbl   <- dataset[ foto_mes<=202105 & foto_mes>=202010,
                    list( "promedio" = mean( get(campo), na.rm=TRUE)) ,
                    foto_mes ]

  ceros <- dataset[ foto_mes<=202105 ,
                    list( "zero_ratio" = sum(get(campo)==0, na.rm=TRUE)/.N ) ,
                    foto_mes ]
  
  plot(x= 1:nrow(tbl),
       y= tbl$promedio,
       type= "o",
       main= paste0("Promedios  -  ",  campo),
       xlab= "Periodo",
       ylab= "Promedio",
       xaxt= "n"
     )

  axis(1, at=1:nrow(tbl), labels=tbl$foto_mes)

  abline(v=c(1,13,25), col=c("green","green","green"), lty=c(1,1,1), lwd=c(1,1,1))
  
  for( i in 1:nrow(tbl) )
  {
    if( ceros[ i, zero_ratio]> 0.99 &  median(ceros[ , zero_ratio]) < 0.99 )
    {
      abline(v=c(i), col=c("red"), lty=c(1), lwd=c(1))
    }
  }
}

dev.off()

pdf("promedios_nocero_desde202010.pdf")

for( campo in  campos_buenos )
{
  tbl   <- dataset[ foto_mes<=202105 & get(campo)!=0,
           list( "promedio" = mean( get(campo), na.rm=TRUE)) ,
           foto_mes ]

  ceros <- dataset[ foto_mes<=202105 ,
                    list( "zero_ratio" = sum(get(campo)==0, na.rm=TRUE)/.N ) ,
                    foto_mes ]
  
  plot(x= 1:nrow(tbl),
       y= tbl$promedio,
       type= "o",
       main= paste0("Promedios NO cero -  ",  campo),
       xlab= "Periodo",
       ylab= "Promedio valores no cero",
       xaxt= "n"
     )

  axis(1, at=1:nrow(tbl), labels=tbl$foto_mes)

  abline(v=c(1,13,25), col=c("green","green","green"), lty=c(1,1,1), lwd=c(1,1,1))
  
  for( i in 1:nrow(tbl) )
  {
    if( ceros[ i, zero_ratio]> 0.99 &  median(ceros[ , zero_ratio]) < 0.99 )
    {
      abline(v=c(i), col=c("red"), lty=c(1), lwd=c(1))
    }
  }
}

dev.off()

#dejo la marca final
cat( format(Sys.time(), "%Y%m%d %H%M%S"),"\n",
     file= "zRend.txt",
     append= TRUE  )
